The Longing now with the Kiloning!

+++ src/main/java/dev/louis/nebula/spell/source/BlockEntitySpellSource.java
@@ -62,7 +62,7 @@ public class BlockEntitySpellSource<BE extends BlockEntity> implements SpellSour
     }
 
     @Override
-    public boolean drainMana(int amount, TransactionContext context) {
+    public boolean drainMana(long amount, TransactionContext context) {
         return getManaPool().map(manaPool -> (manaPool.extractMana(amount, context) == amount)).orElse(false);
     }
 }
